pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "etcd3"
description = "Python client for the etcd3 API"
dynamic = ["version"]
requires-python = ">=3.10"
readme = "README.rst"
authors = [
  {email = "Louis Taylor <louis@kragniz.eu>"},
  {email = "Ledger SAS <team-embedded-software+enclave@ledger.fr>"},
]
dependencies = [
    "grpcio>=1.76.0rc1",
    "protobuf>=6.33",
    "six==1.17.0",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
tests = [
    "pifpaf",
    "PyYAML>=5.1",
    "bumpversion>=0.5.3",
    "coverage",
    "flake8-import-order",
    "flake8",
    "grpcio-tools>=1.76.0rc1",
    "hypothesis",
    "pytest>=4.6.5",
    "pytest-cov",
    "flake8-docstrings>=1.3.0",
    "mock>=2.0.0",
    "tenacity>=5.0.2",
]
protoc = [
    "grpcio-tools",
    "grpcio>=1.76.0rc1",
]
doc = [
    "Sphinx>=1.8.2",
]

[project.scripts]
genproto = "_genproto:main"

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.13"
show_error_codes = true
check_untyped_defs = true
no_implicit_reexport = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_untyped_imports = true

[tool.bandit.assert_used]
skips = ["tests/*.py"]

[tool.pydocstyle]
convention = "pep257"

[tool.setuptools]
py-modules = ["_genproto"]
